Simplify by merging HorizontalBusLaneManager implementations into PositionFinder implementations
