+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/layout/positionprocessor/AbstractPositionFinder.java
@@ -19,7 +19,7 @@ import java.util.Map;
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public abstract class AbstractPositionFinder implements PositionFinder {
+public abstract class AbstractPositionFinder implements PositionFinder, HorizontalBusListManager {
 
     public abstract Map<BusNode, Integer> indexBusPosition(List<BusNode> busNodes, List<BusCell> busCells);
 
@@ -44,5 +44,14 @@ public abstract class AbstractPositionFinder implements PositionFinder {
 
     public void organizeDirections(VoltageLevelGraph graph, List<Subsection> subsections) {
         forceSameOrientationForShuntedCell(graph);
+    }
+
+    public void mergeHblWithNoLink(BSCluster leftCluster, BSCluster rightCluster) {
+        rightCluster.getHorizontalBusLists().forEach(hbl -> {
+            hbl.shift(leftCluster.getLength());
+            hbl.setBsCluster(leftCluster);
+        });
+        leftCluster.getHorizontalBusLists().addAll(rightCluster.getHorizontalBusLists());
+        rightCluster.getHorizontalBusLists().removeAll(rightCluster.getHorizontalBusLists());
     }
 }
