+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/layout/positionprocessor/HorizontalBusListManager.java
@@ -14,12 +14,4 @@ public interface HorizontalBusListManager {
 
     void mergeHbl(BSCluster leftCluster, BSCluster rightCluster);
 
-    default void mergeHblWithNoLink(BSCluster leftCluster, BSCluster rightCluster) {
-        rightCluster.getHorizontalBusLists().forEach(hbl -> {
-            hbl.shift(leftCluster.getLength());
-            hbl.setBsCluster(leftCluster);
-        });
-        leftCluster.getHorizontalBusLists().addAll(rightCluster.getHorizontalBusLists());
-        rightCluster.getHorizontalBusLists().removeAll(rightCluster.getHorizontalBusLists());
-    }
 }
